Remove development mailing list link from issue template
GitHub now has better facilities for handling users using issues to request features or start a discussion, than it did when the template was originally written. For example we can transfer issues to another repo or convert them into discussions. As a result it is less necessary to tell users not to go elsewhere to request features or start discussions.