Add NavigationObserver utility to observe client-side navigations
This utility monitors for same-document navigations that update the path
or query string of the URL but don't load a new Document.

This uses the Navigation API if available (Chrome >= 102) or falls back
to a combination of the "popstate" event and monkey-patching the
`window.history` API otherwise (all other browsers).