Implement plumbing for new content info banner
Prepare for a new design for the content information banner that is used in
JSTOR PDF assignments by:

 - Replacing the `contentPartner` configuration string with a `contentInfoBanner`
   object.
 - Replacing the `ContentPartnerBanner` component with a `ContentInfoBanner`
   component that has a placeholder design, using data from the
   `contentInfoBanner` configuration.

The `contentInfoBanner` configuration contains metadata about the document being
displayed, a logo for the content provider, and links to the previous/next items
in whatever collection the document belongs to. There are no JSTOR-specific
details hardcoded, which should enable this banner to be used with other content
sources in future, if required.

The `contentInfoBanner` configuration can be supplied either on the initial
launch or, in future, at some later point. The display of the banner is
initiated by the Guest calling a `showContentInfo` method on the integration.
This method is currently only implemented for the PDF integration.