Transmit content info banner data from sidebar to main guest
When data for the content info banner has been received by the sidebar and added
to the store, send it to the main guest in a `showContentInfo` call. The guest
then calls the integration's `showContentInfo` method to actually display the
data to the user, in a location that is appropriate for the content type.