Send content banner data to all guests
Even though this data is only used by (or at least, we expect it to only be used
by) the main guest, we can get it there faster if we send the data to all
guests. This avoids the need to wait for the `documentInfoChanged` event to be
sent by the guest to the sidebar, in order for the sidebar to know whether a
guest is the main one or not. In the context of PDFs, this event is only sent
once the full PDF has been loaded.

If in future we do need to distinguish between main and non-main guests when
sending this event, it would be better to rework the guest <-> sidebar
connection protocol so that the sidebar can know immediately whether a guest is
the main one or a sub-frame, without having to wait for `documentInfoChanged`.

As a bonus, sending the data to all guests simplifies some tests.