Convert state of drafts and frames store modules to objects
Change the state of the drafts and frames store modules from arrays to objects
containing a single array-valued field. This will allow additional fields to be
added to these modules in future. It also means that store modules are now more
uniform, as all modules now have an object, rather than array, for their
`initialState`.