Fix alignment of replies in RTL languages
`<li>` elements have a default style of `text-align: match-parent`. This
resulted in a computed `text-align: left` style for paragraphs in replies, which
have an `<li>` ancestor. As a result replies written in RTL languages (eg.
Hebrew) were rendered left-to-right, instead of right-to-left as intended. Fix
this by adding a `text-align: start` on the `<li>`. This sets the alignment for
descendants to match the computed `direction` property. The `direction` for
replies is computed based on their content due to a `dir="auto"` attribute on
`StyledText` components.

Fixes https://github.com/hypothesis/client/issues/4705