Ensure `<hypothesis-*>` elements do not clip their contents
Work around an issue on OpenStax pages where a `* { overflow: auto }` rule
causes `<hypothesis-*>` elements to clip their contents. This issue prevented
the adder toolbar from being shown when text was selected.

I chose to use a `:host` rule here so that the rule could be colocated in the
same stylesheet as the rest of the styles related to these elements. However
this does mean that `!important` needs to be used for any rule that might
conflict with a host page rule. See Stack Overflow links in comments.

Fixes #4641