Remove polyfill loader from the boot script
None of the existing polyfills are used in any supported browsers, and going
forwards we have tentatively agreed [1] to use regular imports for polyfills,
instead of this extra machinery.

This should make it easier to see where polyfills have come from,
and the worse ergonomics (eg. of having to add an extra import, calling a
standalone function instead of a method) haven't proven a major issue.

[1] See discussion on https://github.com/hypothesis/client/pull/4722