Use "hovered" instead of "focused" to refer to hovered annotations/highlights
The term "focused" is overloaded as it can refer to several different
states in which an annotation can be specially marked:

 1. When the user hovers an annotation card or the corresponding
    highlights in the document or buckets in the bucket bar

 2. When the annotation card has keyboard focus

 3. When there is a focused user (eg. when grading a student in the LMS
    app) and the annotation belongs to that user

To try and reduce this confusion, use the term "hovered" to refer to
(1). The `focusAnnotations` methods and messages have been renamed
`hoverAnnotations`, which is not strictly accurate as it is the user who
is doing the hovering, but I think it will make sense.