Make ThreadCard programatically focusable only and style focus ring
We need to set a `tabIndex` so that the card can be focused programatically when
the user clicks the "Show" button on the adder toolbar. However we're not sure
if we want the card to appear in the keyboard tab focus sequence yet, so set the
value to -1 for now.

Add the focus-visible-ring class so that the card gets out standard focus ring
when it is focused.