Implement browser extension update trigger after release
Fill in the steps to trigger the update-client workflow in the
browser-extension repository after a new client version is released.

The GITHUB_TOKEN that actions get by default is scoped to the current
repo, so a different one is required for cross-repository workflow
triggers. There was a discussion around which user or app the token
should be associated with, see [1].

[1] https://hypothes-is.slack.com/archives/C4K6M7P5E/p1663845478224189