Use default environment for update-extension job
Using the "production" environment for the job results in a prompt for
approval, which we only want to happen for the actual production
release. Using the default environment should allow this job to execute
automatically after the production client release.

The secret used by this job has been migrated from being an environment
secret to a repository secret.