Remove CI status badge from README
This badge is no longer working because the CI workflow does not run
directly on the `main` branch. Instead the Release workflow runs, and CI
is part of that. We could make the badge reference this workflow
instead, but the badge is now somewhat redundant as check statuses are
easily available from the Actions tab on GitHub. This is different than
when badges originally became popular and only third-party CI (eg.
Travis) was available.