Update shared components in `MenuArrow`, `AutocompleteList`
As `MenuArrow` now uses two icons — `PointerUpIcon` and
`PointerDownIcon` — instead of a CSS transform (rotate 180 degrees)
for the down variant, absolute positioning had to be adjusted in the
components that use this arrow.

A future enhancement might be to make positioning easier/unnecessary for
components using `MenuArrow`. Some additional comments have been added
for now.

The shared components in `AutocompleteList` are also updated here.