Extract code for `<mosaic-book>` element into a separate module
As part of the work to treat VitalSource books as a single document, the
VitalSource integration will use additional APIs exposed by the `<mosaic-book>`
element which will need to be mocked. Extract the existing implementation out of
vitalsource-epub.mustache so it can in future be re-used by the mock VitalSource
PDF book, and also because a standalone JS file is easier to apply tools to (eg.
Prettier).