Convert shared annotator/ types to TypeScript
This is a purely syntactic change. The documentation and types are the same as
before. One thing to note about using TS syntax is that a module only exports
types which are explicitly marked for export, whereas in JS any types defined in
the module or imported from elsewhere are exported.