Fix unwanted lint errors about type definition order in types/annotator.ts
The TypeScript compiler does not require type definitions to come before their
use, so this is a stylistic preference.

Per the docs [1], it is necessary to first disable ESLint's no-use-before-define
rule to avoid incorrect warnings in TS code, and then enable the corresponding
TS version with `ignoreTypeReferences` and `typedefs` set as desired.

[1] https://typescript-eslint.io/rules/no-use-before-define