Use book or chapter URL in VS integration depending on "book_as_single_document" flag
Add the plumbing needed to make the VitalSource integration return either the
book or chapter URL depending on the value of the "book_as_single_document"
flag. Additionally handle changes in this flag after the integration is created
by notifying the guest about the URI change.

When the flag is enabled, the "book URL" is currently a placeholder that will
in future be replaced by a real value fetched from the `<mosaic-book>`
element in the container frame.