Add EPUBContentSelector and PageSelector selectors to VS annotations
These will be used for various purposes described in the previous commit.
EPUBContentSelector selectors are captured for all VS books.  PageSelector
selectors are currently only captured for PDF-based books where we know the page
numbers are "fixed" and should not change depending on viewer settings. We
_might_ be able to capture them for EPUBs as well in future, though we need to
understand more about how reliable/future-proof they are in that context.