Add functions for comparing CFIs and stripping assertions
When sorting EPUB annotations by document location, we will need to consider
both the chapter/page and the text position within the chapter/page. Add
utilities for comparing CFIs that will be useful for this purpose.

The full sorting rules are specified by
https://idpf.org/epub/linking/cfi/#sec-sorting. These utilities currently only
implement the steps needed to compare simple CFIs which identify a location in
the book's spine (eg. "/2/4/6"), rather than a "full" CFI which also contains a
step indrection ("!") followed by the location of the selected content in the
chapter.