Add `$cluster` client property to annotation objects
This property is set when an annotation object is initialized in:

* annotator: In `guest`, when a user creates a new annotation or highlight
  via the adder controls
* sidebar: In the `annotations` store module when annotation objects are
  initialized before being added to the store

This property is communicated between the sidebar and the annotator when
exchanging annotation data. In the annotator, the value of this property
is used to set an additional CSS class on drawn anchor highlights
(`<hypothesis-highlight>` `span`s).

The presence of this CSS class will allow subsequent differentiated styling
for highlight clusters.