Add additional tests and clarify string vs number step comparisons
CFI steps that cannot be parsed as base-10 integers are invalid, but we try to
fail gracefully in `compareCFIs` by using a string comparison.  This introduces
a problem when one step is an integer and another is a string. Handle this by
arbitrarily deciding that numbers sort before strings. We could have chosen
other approaches as well.