Pass information about current document segment to sidebar
Enable the sidebar to know which segment of a document (eg. which chapter in an
EPUB) is loaded in a guest frame. The sidebar will then be able to use this to
know which annotations associated with the document it should send to the guest,
as well as whether to trigger a segment/chapter navigation when scrolling to an
annotation.