Only send annotations to corresponding chapters in EPUBs
When new annotations are loaded in the store, compare their segment information
(ie. which EPUB chapter they were made in) against that of the guest frames. If
there is no match, skip sending the annotation to the frame. This prevents
annotations made on chapters other than the current one from mis-anchoring
or being incorrectly marked as orphans.

Currently these "skipped" annotations will appear only in the sidebar after a
short delay. This is because logic elsewhere in the application expects all
loaded annotations to be marked as anchored or not and doesn't display them
until the anchoring status is set, or a 500ms timeout is reached. For these
skipped annotations, they don't appear until after the 500ms timeout expires.
This will be addressed separately.