Stop polling for navigations once a frame is disconnected
Fix a bug in `onDocumentReady` where the callback would fire continuously with
an error after the frame either navigated to a cross-origin URL, or was removed
from its parent frame. Also emit a more appropriate error message in the case
where the frame is disconnected.

In the VitalSource integration this issue manifested in an `onDocumentReady`
callback firing continuously for the previous chapter's iframe when it was
removed from the document. This didn't cause an error because the callback
registered by the VS integration does nothing in that case, but it wasted
resources.