Make temporary VitalSource page loaded after navigation more realistic
Make the temporary page loaded during a VitalSource chapter navigation
more realistic by:

 1. Having a URL that is not `about:blank`.
 2. Having content with a `#page-content` element

Change (1) enables the `onDocumentReady` helper to realize that the initial
`about:blank` URL of the new chapter's iframe is about to navigate to a
different URL (the URL of the temporary page). See the
`hasBlankDocumentThatWillNavigate` function.

Change (2) allows the `onDocumentReady` callback in `VitalSourceInjector` to
detect that the temporary page is not the real chapter content, and skip loading
the client into it.

Together these two changes are steps towards more seamless handling of chapter
transitions in the sidebar.