Use same frame ID for each VitalSource content frame
Use the same "vitalsource-content" frame ID for each VitalSource guest
frame. When a chapter navigation occurs, this will enable the sidebar to
associate the new guest, for the new chapter, with the `Frame` object in
the store for the previous guest.

This also gives the frame a more helpful human-readable ID for debugging
purposes.