Wait for feature flag state before sending document info in VitalSource
In order to facilitate a seamless transition after a chapter navigation in
VitalSource, it will be necessary for the Guest in the new frame to report
document information based on the same `book_as_single_document` feature flag
state as for the previous frame.

Allow integrations to signal whether the Guest should wait to receive feature
flags from the sidebar before sending document info via `documentInfoChanged`,
and enable this for VitalSource. We don't enable this for all integrations
because it adds latency on the critical path to fetching and displaying
annotations.