Update documentation for FrameSyncService
Rewrite the description of this service to state that annotations are
synchronized between the sidebar and _guest_ frames, not the host. Also note
some (but not all) of the service's additional responsibilities. In the process
it becomes clear that the service is tending towards being a little overloaded
and it may make sense to split up some of these roles in future.