Handle hovering annotations which are not loaded in guest frame
When a user hovers an annotation card for an EPUB chapter that is different than
the one currently loaded in the guest, and then clicks on that annotation to
navigate the guest to that chapter, ensure that the highlights in the document
are displayed in a hovered state once anchoring completes.

This is handled similar to `FrameSyncService.scrollToAnnotation`, by setting a
flag if the annotation is not loaded in the guest, which triggers a
"hoverAnnotations" RPC call once the annotation is anchored.