Display chapter headings above cards in sidebar
Visually group ebook annotations by chapter in the sidebar by displaying
chapter headings above groups of annotations from a particular chapter.

For each thread a "heading key" is extracted, which is currently the EPUB
Content Document's CFI, taken from the "EPUBContentSelector" selector. For other
annotation types we could use a different key in future. When rendering threads
a heading is displayed above each thread where the key is different than the
previously rendered thread.

To avoid adding complexity to the virtualization calculations, the headings are
rendered as part of the first thread in the group, and so the height of the
heading element is included in the measured height for that thread.