Assume annotations for other document segments will anchor
When annotations are loaded into the sidebar, they are not displayed in the
"Annotations" tab immediately. Instead the client waits until the guest reports
that the annotation was successfully anchored, or a timeout expires. This is to
avoid annotations briefly appearing in the Annotations tab and then "jumping" to
the Orphans tab a moment later if they fail to anchor.

For annotations in document segments (eg. EPUB chapters) other than the one
currently loaded in a guest frame, the annotation will never anchor in the
guest, and so the annotation did not appear in the sidebar until the anchoring
timeout expired. The result was that current-chapter annotations appeared
quickly, and other-chapter annotations appeared in the sidebar several hundred
ms later.

This commit improves the experience by optimisitically assuming annotations
in other chapters will anchor when the user navigates to those chapters, and
immediately marking them as "anchored" in the store. If they fail to anchor
when the user navigates to the chapter, then the status will be updated and
the annotation will be marked as an orphan at that point.