Style highlight clusters with defined colors instead of CSS blending
Set explicit second and third colors to use for nested highlights in
highlight clusters, instead of using `mix-blend-mode: multiply`.

This will allow us to style drawn SVG highlights for PDFS, as transparency
or blending is not feasible for those.

Reorganize CSS and CSS variables a bit to prepare for next steps.