Unload client from iframes when it is unloaded in the host frame
When the client is unloaded in a frame H, ensure that it is also unloaded from
guest frames whose host frame is H. This enables toggling the browser
extension's active state to work properly in VitalSource Bookshelf and other
pages which have annotation-enabled iframes.

This works by having the `Guest` instance in the iframe listen for the existing
"close" message from its connection to the host frame, and relaying that to the
entry point via a "hostDisconnected" event, which then handles it in the same
way as if the client had been unloaded directly in the guest frame.