Remove temporary client configuration from frame when client is unloaded
When the client is unloaded in a guest-only iframe, remove the client
configuration that was added by the `injectClient` function. This is
necessary for subsequent reloading of the client to work, as
`injectClient` aborts if the configuration already exists in a frame.

This change allows the client to be properly de-activated and then later
re-activated in a VitalSource book.