Add temporary fix for crash when creating page notes
The `getSegmentSelector` function assumed all annotations have at least one
target. This is true for all saved annotations and all new annotations with
selectors, but is not true for new page notes.

This is a stop-gap fix. The better solution will be to eliminate the
inconsistency and make it so that all annotations have a non-empty `target`
field.