WIP: Handle cases when text node corresponding to the end of a range has no selected text in it
On certain PDFs, selections that include the last character(s) on a page were not annotatable. This is because the `Range` generated for these selections started and ended in different text layer containers. PDFjs generates one text layer per page, so our code interprets ranges that start in one text layer and end in another as spanning two pages — which we do not support.

However, in these cases, there is no actual text selected in any of the end container’s text nodes.

Adjust range start/end-resolving logic to handle cases in which the range’s end container differs from the start container, but the node corresponding to the `end` node has no selected text in it (i.e. `offset` of `0`). In these cases, resolve backwards to the nearest previous text node.