Add utility to trim text-content Ranges
Remove any leading or trailing whitespace and/or nodes from the supplied
Range. The resulting range starts exactly at the first non-whitespace
character in the Range and ends exactly at the last non-whitespace
character in the Range.

`trimRange` assumes the provided Range starts and ends on a text node
and will throw otherwise.