Add `getAnnotatableRange` to IntegrationBase interface and classes
Update annotator `IntegrationBase` interface to change `canAnnotate` to
`getAnnotatableContent`. This gives the integration an opportunity to
adjust the Range as needed (e.g. trim it) as part of validating whether
it contains any annotatable content.

Update `Guest` to check the `getAnnotatableContent` results for its
integration when a user selects content.