Always capture EPUBContentSelector, PageSelector in VitalSource books
In preparation for enabling the `book_as_single_document` feature for everyone,
enable capturing the EPUBContentSelector selector whether the feature flag is
enabled or not.

Once this is released, all new VS annotations will have all the data they will
need after they are migrated to the new format [1] and only the annotation URL
will need to be changed. This will leave us with only a fixed set of older
annotations for which we will need to obtain the missing CFI and chapter title
data.

[1] See https://github.com/hypothesis/h/issues/7709