Update link to WAI-ARIA combobox example
The previous link is no longer extant (redirects to the landing page for
the authoring guide). The current implementation supports the documented
navigation and behavior guidelines in that example, except where not
applicable (e.g. there is no way to explicitly open the listbox if it is
empty). These types of allowed exceptions are alluded to on
https://www.w3.org/WAI/ARIA/apg/patterns/combobox/