AnontationShareControl: get group from store instead of prop
Refactor to relieve a long-standing FIXME in the
`AnnotationShareControl` component, which pre-dates the completion of
the conversion of the app's UI to preact and the full complement of
available store modules.

As groups and annotations are loaded separately, we cannot guarantee
that group metadata is loaded for a given annotation's group (ID).

Attempt to retrieve the annotation's group from the store, and if it
is undefined, don't render the component.

See https://github.com/hypothesis/client/issues/1542