Avoid using named text-size classes to set specific icon dimensions
In some cases icons are sized relative to local text or containers
(em-based sizing) and that's fine, but using a named text-size
utility directly on an icon can add a layer of unhelpful abstraction if
our intent is simply to size the icon to particular pixel dimensions. It
also entangles icon sizing with text sizing in an unintended way.

In cases where we want to size icons to specific dimensions,
irrespective of local text size, use size utility classes, not text-size
utility classes.

In cases where icons use em-based sizing (e.g. we want to set a text
size on a container such that descendant icons, which are sized at 1em,
have desired dimensions), use arbitrary text sizing classes (e.g.
`text-[16px]` instead of `text-lg`) to be explicit about the intent.