Remove the `tiny` text size token from tailwind
There are a few places in the app where we want to size things to 10px,
typically for teeny-tiny icons, but also in one case, some `<sup>` text.

This isn't really a text size, in terms of the relative text-size scale.
Remove the token from the Tailwind config and use a direct pixel size in
the one remaining usage.