Update `MenuItem` to take `IconComponent`s, `leftChannelContent`
Previously, `MenuItem`'s `icon` prop accepted a string, which could be:

- the name of a registered icon
- a URL (in which case an `img` with that URL as `src` was rendered)
- the string "blank" (in which case space was added at left)

This became over-complex when migrating to use updated icon components.
Now the `icon` prop expects an `IconComponent`, but `MenuItem` also
takes a `leftChannelContent` prop. This allows consumers to set any
content in the left channel — e.g. an image instead of an icon — and
reduces the complexity of this component.

`GroupListItem` has been updated to set `leftChannelContent` as needed
instead of passing "blank" or a URL to an image as the `icon` prop.

Other components updated to provide `IconComponent`s instead of strings.