Fail loudly if VitalSource page info fields are missing
If for any reason VitalSource's `<mosaic-book>` API does not return the page
info fields that the Hypothesis client is expecting, fail loudly so we find out
about the issue, rather than silently creating annotations with
missing/incorrect chapter location information.

I don't currently know of any situation where this is expected to happen, so
this is protection against future changes or differences in behaviors with
certain books that we have not observed yet.