Prevent focus from leaking and inadvertently closing the panel
In context, `AnnotationShareControl` is rendered within its associated
`ThreadCard`. `ThreadCard`s are focusable, so any click within the
panel that is not on a focusable element will "leak out" and focus the
annotation card. This will cause the panel to close because the focus
event is outside of the panel.

Prevent focus from leaking by making the panel itself focusable.

Fixes https://github.com/hypothesis/client/issues/5196