Add PDF.js example document to test set
Add a LaTeX-generated PDF to the example set. This is the same PDF that is used
on PDF.js's demo site at https://mozilla.github.io/pdf.js/web/viewer.html.

It differs from existing test papers in that it includes a two-column layout,
and content that is a mixture of figures, tables, and body text. It also has
different fonts for body text, math and code.