Remove local `tall` media query (screen) and consolidate filter height
The `tall` `screen` value in this project's config conflicts with the
`tall` value in the `frontend-shared` Tailwind preset. This project's
`tall` was only being used in one place, and it turns out the styling
needs there can be more cleanly accomplished without media queries.

Use `min()` to set a constrained max-height on filter-select lists.