Propagate `profileAppUrl` config to guest iframes
Propagate the `profileAppUrl` config setting from the initial frame where
Hypothesis is loaded to guest frames. The boot script expects this setting in
all non-sidebar frames, along with `sidebarAppUrl` and `notebookAppUrl`, because
it doesn't differentiate between host and guest frames when loading resources.
Ideally we would fix that, but it requires a more substantial refactoring that
I don't want to do just now.