Enable host page to respond async to `scrolltorange` event
Add a `waitUntil` method to the `scrolltorange` event which allows the host page
to signal to the client that it should wait for async work to complete before
the client attempts to scroll to a highlight. This is useful in web applications
which need to perform an async re-render of the UI before the content can be
scrolled to.

The initial use case is enabling Via's video player app to clear transcript
search filters and re-render the Preact UI before scrolling to a highlight [1]

[1] See https://github.com/hypothesis/via/pull/932.