Add a note about a scenario where the thresholds will be inappropriate
This function is used to scroll to the "nearest" highlight when clicking the
top/bottom buckets in the bucket bar. This can likely be reimplemented in a
simpler way in future that doesn't rely on knowing where the buckets are
relative to the viewport. If instead we assume that all the passed-in anchors
are off-screen, we could just scroll to the top/bottom one.