Revise `pointer-events` handling for bucket bar container
In 34521471252d40f706afbbf63a92fc8ab7f7a3aa the `pointer-events: none` style was
removed from the bucket bar background. This fixed a problem where clicks
in-between buckets (eg. when trying to click on a bucket) could cause the
sidebar to close. It introduced a new issue however where it made clicking on
page elements behind the bucket bar impossible. This can interfere with clicking
on scroll bar thumbs in pages where the thumb gets drawn below the Hypothesis
sidebar for example.

This commit adds `pointer-events: none` back to this container. The next commit
will address the problem that removing it aimed to resolve, but in a different
way.